+++ src/main/java/entity/nonDynamic/StaticEntity.java
@@ -7,7 +7,7 @@ public abstract class StaticEntity extends Entity {
 
 	public Level level;
 
-	protected StaticEntity(int x, int y, int z, Level level) {
+	protected StaticEntity(float x, float y, int z, Level level) {
 		super(x, y, z);
 		this.level = level;
 	}
